+++ packages/tools/src/init.ts
@@ -5,12 +5,13 @@ import { Events as TOOLS_EVENTS } from './enums';
 import { addEnabledElement, removeEnabledElement } from './store';
 import { resetCornerstoneToolsState } from './store/state';
 import {
+  annotationCompletedListener,
   annotationSelectionListener,
+  annotationModifiedListener,
   segmentationDataModifiedEventListener,
   segmentationRepresentationModifiedEventListener,
   segmentationRepresentationRemovedEventListener,
   segmentationModifiedListener,
-  annotationModifiedListener,
 } from './eventListeners';
 import { annotationInterpolationEventDispatcher } from './eventDispatchers';
 
@@ -104,6 +105,11 @@ function _addCornerstoneToolsEventListeners() {
    * Annotation
    */
   eventTarget.addEventListener(
+    TOOLS_EVENTS.ANNOTATION_COMPLETED,
+    annotationCompletedListener
+  );
+
+  eventTarget.addEventListener(
     TOOLS_EVENTS.ANNOTATION_MODIFIED,
     annotationModifiedListener
   );
@@ -148,6 +154,11 @@ function _removeCornerstoneToolsEventListeners() {
   /**
    * Annotation
    */
+  eventTarget.removeEventListener(
+    TOOLS_EVENTS.ANNOTATION_COMPLETED,
+    annotationCompletedListener
+  );
+
   eventTarget.removeEventListener(
     TOOLS_EVENTS.ANNOTATION_MODIFIED,
     annotationModifiedListener