+++ packages/tools/src/utilities/contourSegmentation/index.ts
@@ -1,0 +1,2 @@
+export { default as areSameSegment } from './areSameSegment';
+export { default as isContourSegmentationAnnotation } from './isContourSegmentationAnnotation';